refactor: move route macro machinery into route/ module Closes #744
