+++ rapina-macros/src/lib.rs
@@ -1,73 +1,12 @@
-use heck::ToKebabCase;
 use proc_macro::TokenStream;
 use quote::quote;
 use syn::spanned::Spanned;
 use syn::{FnArg, ItemFn, LitStr, Pat};
 
-/// Parsed route macro attribute: `"/path"`, `"/path", group = "/prefix"`,
-/// `"/path", description = "..."`, or any combination thereof.
-struct RouteAttr {
-    path: LitStr,
-    group: Option<LitStr>,
-    description: Option<LitStr>,
-}
-
-impl syn::parse::Parse for RouteAttr {
-    fn parse(input: syn::parse::ParseStream) -> syn::Result<Self> {
-        let path: LitStr = input.parse()?;
-        let mut group: Option<LitStr> = None;
-        let mut description: Option<LitStr> = None;
-
-        while input.peek(syn::Token![,]) {
-            input.parse::<syn::Token![,]>()?;
-            if input.is_empty() {
-                break;
-            }
-            let ident: syn::Ident = input.parse()?;
-            input.parse::<syn::Token![=]>()?;
-            if ident == "group" {
-                let value: LitStr = input.parse()?;
-                group = Some(value);
-            } else if ident == "description" {
-                let value: LitStr = input.parse()?;
-                description = Some(value);
-            } else {
-                return Err(syn::Error::new(
-                    ident.span(),
-                    "expected `group` or `description`",
-                ));
-            }
-        }
-
-        if !input.is_empty() {
-            return Err(input.error("unexpected tokens after route attribute"));
-        }
-        Ok(RouteAttr {
-            path,
-            group,
-            description,
-        })
-    }
-}
-
-/// Join a group prefix with a route path at compile time.
-fn join_paths(prefix: &str, path: &str) -> String {
-    let prefix = prefix.trim_end_matches('/');
-    if path.is_empty() || path == "/" {
-        if prefix.is_empty() {
-            return "/".to_string();
-        }
-        return prefix.to_string();
-    }
-    let path = if path.starts_with('/') {
-        path.to_string()
-    } else {
-        format!("/{path}")
-    };
-    format!("{prefix}{path}")
-}
-
+mod route;
 mod schema;
+
+use route::route_macro;
 
 /// Registers a GET route handler.
 ///
@@ -175,695 +114,6 @@ pub fn public(_attr: TokenStream, item: TokenStream) -> TokenStream {
         }
     }
     .into()
-}
-
-fn route_macro_core(
-    method: &str,
-    attr: proc_macro2::TokenStream,
-    item: proc_macro2::TokenStream,
-) -> proc_macro2::TokenStream {
-    let route_attr: RouteAttr = syn::parse2(attr).expect("expected path as string literal");
-    let path_str = if let Some(ref group) = route_attr.group {
-        let g = group.value();
-        assert!(
-            g.starts_with('/'),
-            "group prefix must start with `/`, got: {g:?}"
-        );
-        join_paths(&g, &route_attr.path.value())
-    } else {
-        route_attr.path.value()
-    };
-    let mut func: ItemFn = syn::parse2(item).expect("expected function");
-
-    let func_name = &func.sig.ident;
-    let func_name_str = func_name.to_string();
-    let func_vis = &func.vis;
-
-    // Extract #[public] attribute if present (when #[public] is below the route macro)
-    let is_public = extract_public_attr(&mut func.attrs);
-
-    // Resolve description: explicit attr wins, then first rustdoc line, then None
-    let description_value: Option<String> = route_attr
-        .description
-        .as_ref()
-        .map(|l| l.value())
-        .or_else(|| extract_doc_description(&func.attrs));
-
-    // Extract #[errors(ErrorType)] attribute if present
-    let error_type = extract_errors_attr(&mut func.attrs);
-
-    // Extract #[cache(ttl = N)] attribute if present
-    let cache_ttl = extract_cache_attr(&mut func.attrs);
-
-    let error_responses_impl = if let Some(err_type) = &error_type {
-        quote! {
-            fn error_responses() -> Vec<rapina::error::ErrorVariant> {
-                <#err_type as rapina::error::DocumentedError>::error_variants()
-            }
-        }
-    } else {
-        quote! {}
-    };
-
-    // Extract return type for schema generation
-    let response_schema_impl = if let syn::ReturnType::Type(_, return_type) = &func.sig.output {
-        if let Some(inner_type) = extract_json_inner_type(return_type) {
-            quote! {
-                fn response_schema() -> Option<serde_json::Value> {
-                    Some(rapina::openapi_schema_for::<#inner_type>())
-                }
-            }
-        } else {
-            quote! {}
-        }
-    } else {
-        quote! {}
-    };
-
-    // Extract request body type and content type for schema generation.
-    // Only generate requestBody for POST, PUT, and PATCH methods per OpenAPI spec.
-    let (request_schema_impl, request_content_type_impl, request_body_required_impl) =
-        if matches!(method, "POST" | "PUT" | "PATCH") {
-            if let Some(meta) = extract_request_body_meta(&func.sig.inputs) {
-                let inner_type = meta.inner_type;
-                let content_type = meta.content_type;
-                let required = meta.required;
-                (
-                    quote! {
-                        fn request_schema() -> Option<serde_json::Value> {
-                            Some(rapina::openapi_schema_for::<#inner_type>())
-                        }
-                    },
-                    quote! {
-                        fn request_content_type() -> Option<&'static str> {
-                            Some(#content_type)
-                        }
-                    },
-                    quote! {
-                        fn request_body_required() -> Option<bool> {
-                            Some(#required)
-                        }
-                    },
-                )
-            } else {
-                (quote! {}, quote! {}, quote! {})
-            }
-        } else {
-            (quote! {}, quote! {}, quote! {})
-        };
-
-    // Collect header params (also strips #[header("name")] attrs from inputs)
-    let header_params = match collect_header_params(&mut func.sig.inputs) {
-        Ok(p) => p,
-        Err(e) => return e.to_compile_error(),
-    };
-
-    // Build an index: arg_idx → &HeaderParamMeta for O(1) lookup during codegen
-    let header_by_arg: std::collections::HashMap<usize, &HeaderParamMeta> =
-        header_params.iter().map(|p| (p.arg_idx, p)).collect();
-
-    // Build header_parameters() impl for the Handler trait
-    let header_parameters_impl = if header_params.is_empty() {
-        quote! {}
-    } else {
-        let entries = header_params.iter().map(|p| {
-            let name = &p.name;
-            let required = p.required;
-            quote! {
-                rapina::discovery::HeaderParamInfo {
-                    name: #name.to_string(),
-                    required: #required,
-                }
-            }
-        });
-        quote! {
-            fn header_parameters() -> Vec<rapina::discovery::HeaderParamInfo> {
-                vec![#(#entries),*]
-            }
-        }
-    };
-
-    // Build description() impl for the Handler trait
-    let description_impl = if let Some(ref desc) = description_value {
-        quote! {
-            fn description() -> Option<&'static str> {
-                Some(#desc)
-            }
-        }
-    } else {
-        quote! {}
-    };
-
-    let args: Vec<_> = func.sig.inputs.iter().collect();
-
-    // Extract return type for type annotation (helps with type inference in async blocks)
-    let return_type_annotation = match &func.sig.output {
-        syn::ReturnType::Type(_, ty) => quote! { : #ty },
-        syn::ReturnType::Default => quote! {},
-    };
-
-    // Optional cache TTL header injection
-    let cache_header_injection = if let Some(ttl) = cache_ttl {
-        let ttl_str = ttl.to_string();
-        quote! {
-            let mut __rapina_response = __rapina_response;
-            __rapina_response.headers_mut().insert(
-                "x-rapina-cache-ttl",
-                rapina::http::HeaderValue::from_static(#ttl_str),
-            );
-        }
-    } else {
-        quote! {}
-    };
-
-    // Build the handler body
-    // Use __rapina_ prefix for internal variables to avoid shadowing user's variables
-    let handler_body = if args.is_empty() {
-        let inner_block = &func.block;
-        quote! {
-            let __rapina_result #return_type_annotation = (async #inner_block).await;
-            let __rapina_response = rapina::response::IntoResponse::into_response(__rapina_result);
-            #cache_header_injection
-            __rapina_response
-        }
-    } else {
-        let inner_block = &func.block;
-
-        // Check if all args are header extractors (so we never need to split req into parts)
-        let all_headers = args.iter().all(|arg| {
-            if let FnArg::Typed(pt) = arg {
-                detect_header_type(&pt.ty).is_some()
-            } else {
-                false
-            }
-        });
-
-        // Check if the single arg is a header type
-        let single_is_header = args.len() == 1
-            && args.first().is_some_and(|arg| {
-                if let FnArg::Typed(pt) = arg {
-                    detect_header_type(&pt.ty).is_some()
-                } else {
-                    false
-                }
-            });
-
-        if args.len() == 1 && !single_is_header {
-            // Single non-header arg: pass request directly to FromRequest
-            let arg = &args[0];
-            if let FnArg::Typed(pat_type) = arg {
-                let pat = &pat_type.pat;
-                let arg_type = &pat_type.ty;
-                let tmp = syn::Ident::new("__rapina_arg_0", proc_macro2::Span::call_site());
-                quote! {
-                    let #tmp = match <#arg_type as rapina::extract::FromRequest>::from_request(__rapina_req, &__rapina_params, &__rapina_state).await {
-                        Ok(v) => v,
-                        Err(e) => return rapina::response::IntoResponse::into_response(e),
-                    };
-                    let #pat = #tmp;
-                    let __rapina_result #return_type_annotation = (async #inner_block).await;
-                    let __rapina_response = rapina::response::IntoResponse::into_response(__rapina_result);
-                    #cache_header_injection
-                    __rapina_response
-                }
-            } else {
-                unreachable!("handler argument must be a typed pattern")
-            }
-        } else if all_headers {
-            // All args are header extractors — extract from parts, no body split needed
-            let mut header_extractions = Vec::new();
-            for (i, arg) in args.iter().enumerate() {
-                if let FnArg::Typed(pat_type) = arg {
-                    let pat = &pat_type.pat;
-                    let tmp = syn::Ident::new(
-                        &format!("__rapina_arg_{}", i),
-                        proc_macro2::Span::call_site(),
-                    );
-                    let meta = header_by_arg.get(&i).expect("all_headers: missing meta");
-                    header_extractions.push(gen_header_extraction(
-                        &meta.inner_type,
-                        meta.required,
-                        &meta.name,
-                        &tmp,
-                    ));
-                    header_extractions.push(quote! { let #pat = #tmp; });
-                }
-            }
-            quote! {
-                let (__rapina_parts, _) = __rapina_req.into_parts();
-                #(#header_extractions)*
-                let __rapina_result #return_type_annotation = (async #inner_block).await;
-                let __rapina_response = rapina::response::IntoResponse::into_response(__rapina_result);
-                #cache_header_injection
-                __rapina_response
-            }
-        } else {
-            // Multiple args: all but last use FromRequestParts (or header extraction), last uses FromRequest
-            let mut parts_extractions = Vec::new();
-
-            for (i, arg) in args[..args.len() - 1].iter().enumerate() {
-                if let FnArg::Typed(pat_type) = arg {
-                    let pat = &pat_type.pat;
-                    let arg_type = &pat_type.ty;
-                    let tmp = syn::Ident::new(
-                        &format!("__rapina_arg_{}", i),
-                        proc_macro2::Span::call_site(),
-                    );
-                    if detect_header_type(arg_type).is_some() {
-                        let meta = header_by_arg.get(&i).expect("mixed: missing meta");
-                        parts_extractions.push(gen_header_extraction(
-                            &meta.inner_type,
-                            meta.required,
-                            &meta.name,
-                            &tmp,
-                        ));
-                        parts_extractions.push(quote! { let #pat = #tmp; });
-                    } else {
-                        parts_extractions.push(quote! {
-                            let #tmp = match <#arg_type as rapina::extract::FromRequestParts>::from_request_parts(&__rapina_parts, &__rapina_params, &__rapina_state).await {
-                                Ok(v) => v,
-                                Err(e) => return rapina::response::IntoResponse::into_response(e),
-                            };
-                            let #pat = #tmp;
-                        });
-                    }
-                }
-            }
-
-            let last_arg = args.last().unwrap();
-            let last_extraction = if let FnArg::Typed(pat_type) = last_arg {
-                let pat = &pat_type.pat;
-                let arg_type = &pat_type.ty;
-                let last_idx = args.len() - 1;
-                let tmp = syn::Ident::new(
-                    &format!("__rapina_arg_{}", last_idx),
-                    proc_macro2::Span::call_site(),
-                );
-                if detect_header_type(arg_type).is_some() {
-                    let meta = header_by_arg
-                        .get(&last_idx)
-                        .expect("last arg: missing meta");
-                    let header_extr =
-                        gen_header_extraction(&meta.inner_type, meta.required, &meta.name, &tmp);
-                    quote! {
-                        #header_extr
-                        let #pat = #tmp;
-                        // Reconstruct the request (body not consumed for header-only last arg)
-                        let _ = __rapina_body;
-                    }
-                } else {
-                    quote! {
-                        let __rapina_req = rapina::http::Request::from_parts(__rapina_parts, __rapina_body);
-                        let #tmp = match <#arg_type as rapina::extract::FromRequest>::from_request(__rapina_req, &__rapina_params, &__rapina_state).await {
-                            Ok(v) => v,
-                            Err(e) => return rapina::response::IntoResponse::into_response(e),
-                        };
-                        let #pat = #tmp;
-                    }
-                }
-            } else {
-                unreachable!("handler argument must be a typed pattern")
-            };
-
-            quote! {
-                let (__rapina_parts, __rapina_body) = __rapina_req.into_parts();
-                #(#parts_extractions)*
-                #last_extraction
-                let __rapina_result #return_type_annotation = (async #inner_block).await;
-                let __rapina_response = rapina::response::IntoResponse::into_response(__rapina_result);
-                #cache_header_injection
-                __rapina_response
-            }
-        }
-    };
-
-    // Build the router method call for the register function
-    let router_method = syn::Ident::new(&method.to_lowercase(), proc_macro2::Span::call_site());
-    let register_fn_name = syn::Ident::new(
-        &format!("__rapina_register_{}", func_name_str),
-        proc_macro2::Span::call_site(),
-    );
-
-    // Generate the struct, Handler impl, and inventory submission
-    quote! {
-        #[derive(Clone, Copy)]
-        #[allow(non_camel_case_types)]
-        #func_vis struct #func_name;
-
-        impl rapina::handler::Handler for #func_name {
-            const NAME: &'static str = #func_name_str;
-
-            #response_schema_impl
-            #request_schema_impl
-            #request_content_type_impl
-            #request_body_required_impl
-            #error_responses_impl
-            #header_parameters_impl
-            #description_impl
-
-            fn call(
-                &self,
-                __rapina_req: rapina::hyper::Request<rapina::hyper::body::Incoming>,
-                __rapina_params: rapina::extract::PathParams,
-                __rapina_state: std::sync::Arc<rapina::state::AppState>,
-            ) -> std::pin::Pin<Box<dyn std::future::Future<Output = rapina::hyper::Response<rapina::response::BoxBody>> + Send>> {
-                Box::pin(async move {
-                    #handler_body
-                })
-            }
-        }
-
-        #[doc(hidden)]
-        fn #register_fn_name(__rapina_router: rapina::router::Router) -> rapina::router::Router {
-            __rapina_router.#router_method(#path_str, #func_name)
-        }
-
-        rapina::inventory::submit! {
-            rapina::discovery::RouteDescriptor {
-                method: #method,
-                path: #path_str,
-                handler_name: #func_name_str,
-                is_public: #is_public,
-                response_schema: <#func_name as rapina::handler::Handler>::response_schema,
-                request_schema: <#func_name as rapina::handler::Handler>::request_schema,
-                request_content_type: <#func_name as rapina::handler::Handler>::request_content_type,
-                request_body_required: <#func_name as rapina::handler::Handler>::request_body_required,
-                error_responses: <#func_name as rapina::handler::Handler>::error_responses,
-                header_parameters: <#func_name as rapina::handler::Handler>::header_parameters,
-                description: <#func_name as rapina::handler::Handler>::description,
-                register: #register_fn_name,
-            }
-        }
-    }
-}
-
-/// Extracts the inner type from Json<T> wrapper for schema generation
-fn extract_json_inner_type(return_type: &syn::Type) -> Option<proc_macro2::TokenStream> {
-    if let syn::Type::Path(type_path) = return_type
-        && let Some(last_segment) = type_path.path.segments.last()
-    {
-        // Direct Json<T>
-        if last_segment.ident == "Json"
-            && let syn::PathArguments::AngleBracketed(args) = &last_segment.arguments
-            && let Some(syn::GenericArgument::Type(inner_type)) = args.args.first()
-        {
-            return Some(quote!(#inner_type));
-        }
-
-        // Result<Json<T>> or Result<Json<T>, E>
-        if last_segment.ident == "Result"
-            && let syn::PathArguments::AngleBracketed(args) = &last_segment.arguments
-            && let Some(syn::GenericArgument::Type(ok_type)) = args.args.first()
-        {
-            return extract_json_inner_type(ok_type);
-        }
-    }
-    None
-}
-
-/// Extracts the request body metadata from handler function arguments.
-/// Supports Json<T>, Form<T>, Validated<Json<T>>, and Validated<Form<T>>.
-fn extract_request_body_meta(
-    inputs: &syn::punctuated::Punctuated<syn::FnArg, syn::Token![,]>,
-) -> Option<RequestBodyMeta> {
-    for arg in inputs.iter() {
-        if let syn::FnArg::Typed(pat_type) = arg {
-            if let Some(meta) = extract_body_inner_type(&pat_type.ty) {
-                return Some(meta);
-            }
-        }
-    }
-    None
-}
-
-/// Information about a request body extractor.
-struct RequestBodyMeta {
-    inner_type: proc_macro2::TokenStream,
-    content_type: &'static str,
-    required: bool,
-}
-
-/// Extracts the inner type and content type from Json<T>, Form<T>, Validated<Json<T>>/Validated<Form<T>>,
-/// or Option<Json<T>>/Option<Form<T>>.
-fn extract_body_inner_type(ty: &syn::Type) -> Option<RequestBodyMeta> {
-    if let syn::Type::Path(type_path) = ty
-        && let Some(last_segment) = type_path.path.segments.last()
-    {
-        // Direct Json<T>
-        if last_segment.ident == "Json"
-            && let syn::PathArguments::AngleBracketed(args) = &last_segment.arguments
-            && let Some(syn::GenericArgument::Type(inner_type)) = args.args.first()
-        {
-            return Some(RequestBodyMeta {
-                inner_type: quote!(#inner_type),
-                content_type: "application/json",
-                required: true,
-            });
-        }
-        // Direct Form<T>
-        if last_segment.ident == "Form"
-            && let syn::PathArguments::AngleBracketed(args) = &last_segment.arguments
-            && let Some(syn::GenericArgument::Type(inner_type)) = args.args.first()
-        {
-            return Some(RequestBodyMeta {
-                inner_type: quote!(#inner_type),
-                content_type: "application/x-www-form-urlencoded",
-                required: true,
-            });
-        }
-        // Validated<Json<T>> or Validated<Form<T>>
-        if last_segment.ident == "Validated"
-            && let syn::PathArguments::AngleBracketed(args) = &last_segment.arguments
-            && let Some(syn::GenericArgument::Type(inner_extractor)) = args.args.first()
-        {
-            return extract_body_inner_type(inner_extractor);
-        }
-        // Option<Json<T>> or Option<Form<T>> - optional request body
-        if last_segment.ident == "Option"
-            && let syn::PathArguments::AngleBracketed(args) = &last_segment.arguments
-            && let Some(syn::GenericArgument::Type(inner_extractor)) = args.args.first()
-        {
-            if let Some(mut meta) = extract_body_inner_type(inner_extractor) {
-                meta.required = false;
-                return Some(meta);
-            }
-        }
-    }
-    None
-}
-
-/// Extract the first non-empty line from `///` doc comments on a function.
-fn extract_doc_description(attrs: &[syn::Attribute]) -> Option<String> {
-    for attr in attrs {
-        if !attr.path().is_ident("doc") {
-            continue;
-        }
-        if let syn::Meta::NameValue(nv) = &attr.meta {
-            if let syn::Expr::Lit(syn::ExprLit {
-                lit: syn::Lit::Str(s),
-                ..
-            }) = &nv.value
-            {
-                let line = s.value();
-                let trimmed = line.trim();
-                if !trimmed.is_empty() {
-                    return Some(trimmed.to_string());
-                }
-            }
-        }
-    }
-    None
-}
-
-/// Extract #[errors(ErrorType)] attribute from function attributes, removing it if found.
-fn extract_errors_attr(attrs: &mut Vec<syn::Attribute>) -> Option<syn::Type> {
-    let idx = attrs
-        .iter()
-        .position(|attr| attr.path().is_ident("errors"))?;
-    let attr = attrs.remove(idx);
-    let err_type: syn::Type = attr.parse_args().expect("expected #[errors(ErrorType)]");
-    Some(err_type)
-}
-
-/// Extract #[cache(ttl = N)] attribute from function attributes, removing it if found.
-fn extract_cache_attr(attrs: &mut Vec<syn::Attribute>) -> Option<u64> {
-    let idx = attrs
-        .iter()
-        .position(|attr| attr.path().is_ident("cache"))?;
-    let attr = attrs.remove(idx);
-
-    let mut ttl: Option<u64> = None;
-    attr.parse_nested_meta(|meta| {
-        if meta.path.is_ident("ttl") {
-            let value = meta.value()?;
-            let lit: syn::LitInt = value.parse()?;
-            ttl = Some(lit.base10_parse()?);
-            Ok(())
-        } else {
-            Err(meta.error("expected `ttl`"))
-        }
-    })
-    .expect("expected #[cache(ttl = N)]");
-
-    ttl
-}
-
-/// Extract #[public] attribute from function attributes, removing it if found.
-fn extract_public_attr(attrs: &mut Vec<syn::Attribute>) -> bool {
-    if let Some(idx) = attrs.iter().position(|attr| attr.path().is_ident("public")) {
-        attrs.remove(idx);
-        true
-    } else {
-        false
-    }
-}
-
-/// Generate the extraction code for a `Header<T>` or `Option<Header<T>>` parameter.
-///
-/// `header_name` is the resolved HTTP header name (kebab-case, possibly from
-/// an explicit `#[header("name")]` attribute).
-fn gen_header_extraction(
-    inner_type: &syn::Type,
-    required: bool,
-    header_name: &str,
-    tmp: &syn::Ident,
-) -> proc_macro2::TokenStream {
-    if required {
-        quote! {
-            let #tmp = match rapina::extract::extract_header::<#inner_type>(&__rapina_parts, #header_name) {
-                Ok(v) => rapina::extract::Header::new(#header_name, v),
-                Err(e) => return rapina::response::IntoResponse::into_response(e),
-            };
-        }
-    } else {
-        quote! {
-            let #tmp = match rapina::extract::extract_optional_header::<#inner_type>(&__rapina_parts, #header_name) {
-                Ok(Some(v)) => Some(rapina::extract::Header::new(#header_name, v)),
-                Ok(None) => None,
-                Err(e) => return rapina::response::IntoResponse::into_response(e),
-            };
-        }
-    }
-}
-
-/// Metadata about a single `Header<T>` or `Option<Header<T>>` parameter.
-struct HeaderParamMeta {
-    /// Zero-based index of this param in the handler's argument list.
-    arg_idx: usize,
-    /// The HTTP header name (e.g. "x-request-id").
-    name: String,
-    /// Whether the parameter is required (`Header<T>`) or optional (`Option<Header<T>>`).
-    required: bool,
-    /// The inner `T` type (for generating the extraction call).
-    inner_type: syn::Type,
-}
-
-/// Extract `#[header("name")]` attribute from a parameter's attribute list.
-///
-/// Returns the explicit header name if present, removing the attribute.
-fn extract_header_attr(attrs: &mut Vec<syn::Attribute>) -> Option<String> {
-    let idx = attrs
-        .iter()
-        .position(|attr| attr.path().is_ident("header"))?;
-    let attr = attrs.remove(idx);
-    let lit: LitStr = attr.parse_args().expect("expected #[header(\"name\")]");
-    Some(lit.value())
-}
-
-/// Detect if `ty` is `Header<T>` (required) or `Option<Header<T>>` (optional).
-///
-/// Returns `Some((inner_type, required))` on match, `None` otherwise.
-///
-/// Matches `Header<T>` (bare or path-qualified as `extract::Header<T>` /
-/// `rapina::extract::Header<T>`).  Any other qualifying path (e.g.
-/// `my_crate::Header<T>`) returns `None`, so user-defined types named `Header`
-/// fall through to normal handling instead of producing a confusing compile
-/// error from macro-generated code.
-fn detect_header_type(ty: &syn::Type) -> Option<(syn::Type, bool)> {
-    let syn::Type::Path(type_path) = ty else {
-        return None;
-    };
-    let last = type_path.path.segments.last()?;
-
-    // Direct Header<T>
-    if last.ident == "Header" {
-        // When the type is qualified (e.g. `foo::Header`), only treat it as
-        // rapina's Header if the leading path is a known rapina prefix.
-        // Bare `Header` (imported via prelude) has no leading segments and
-        // is always accepted.
-        let segments: Vec<_> = type_path.path.segments.iter().collect();
-        let is_rapina_header = match segments.len() {
-            1 => true,                                                            // bare `Header`
-            2 => segments[0].ident == "extract", // `extract::Header`
-            3 => segments[0].ident == "rapina" && segments[1].ident == "extract", // `rapina::extract::Header`
-            _ => false,
-        };
-        if is_rapina_header {
-            if let syn::PathArguments::AngleBracketed(args) = &last.arguments {
-                if let Some(syn::GenericArgument::Type(inner)) = args.args.first() {
-                    return Some((inner.clone(), true));
-                }
-            }
-        }
-    }
-
-    // Option<Header<T>>
-    if last.ident == "Option" {
-        if let syn::PathArguments::AngleBracketed(args) = &last.arguments {
-            if let Some(syn::GenericArgument::Type(inner)) = args.args.first() {
-                if let Some((inner_t, _)) = detect_header_type(inner) {
-                    return Some((inner_t, false));
-                }
-            }
-        }
-    }
-
-    None
-}
-
-/// Collect all `Header<T>` / `Option<Header<T>>` parameters from handler inputs.
-///
-/// Also strips any `#[header("name")]` attributes from the parameters
-/// (they are not valid Rust attributes and must be removed before codegen).
-fn collect_header_params(
-    inputs: &mut syn::punctuated::Punctuated<syn::FnArg, syn::Token![,]>,
-) -> syn::Result<Vec<HeaderParamMeta>> {
-    let mut params = Vec::new();
-
-    for (arg_idx, arg) in inputs.iter_mut().enumerate() {
-        let syn::FnArg::Typed(pat_type) = arg else {
-            continue;
-        };
-
-        let Some((inner_type, required)) = detect_header_type(&pat_type.ty) else {
-            continue;
-        };
-
-        // Check for explicit #[header("name")] override on the parameter
-        let explicit_name = extract_header_attr(&mut pat_type.attrs);
-
-        // Derive header name from snake_case param name, or use explicit override.
-        let name = if let Some(n) = explicit_name {
-            n
-        } else if let Pat::Ident(pat_ident) = &*pat_type.pat {
-            pat_ident.ident.to_string().to_kebab_case()
-        } else {
-            // Destructure pattern — can't infer name, user must use #[header("name")]
-            return Err(syn::Error::new_spanned(
-                &*pat_type.pat,
-                "Header<T> parameter with a destructure pattern must have a #[header(\"name\")] attribute",
-            ));
-        };
-
-        params.push(HeaderParamMeta {
-            arg_idx,
-            name,
-            required,
-            inner_type,
-        });
-    }
-
-    Ok(params)
 }
 
 /// Registers a channel handler for the relay system.
@@ -1373,10 +623,6 @@ fn job_macro_impl(
     }
 }
 
-fn route_macro(method: &str, attr: TokenStream, item: TokenStream) -> TokenStream {
-    route_macro_core(method, attr.into(), item.into()).into()
-}
-
 /// Derive macro for type-safe configuration
 ///
 /// Generates a `from_env()` method that loads configuration from environment variables.
@@ -1536,9 +782,8 @@ fn derive_config_impl(input: proc_macro2::TokenStream) -> proc_macro2::TokenStre
 
 #[cfg(test)]
 mod tests {
-    use super::{
-        job_macro_impl, join_paths, metric_macro_impl, relay_macro_impl, route_macro_core,
-    };
+    use super::{job_macro_impl, metric_macro_impl, relay_macro_impl};
+    use crate::route::{join_paths, route_macro_core};
     use quote::quote;
 
     #[test]
